Added basic ability to edit names of todo items

+++ functions.go
@@ -6,21 +6,56 @@ import (
 	"log"
 	"os"
 	"os/exec"
+	"strconv"
 	"strings"
 )
 
-func list(todos []Item, finished []Item) {
+func list(todos []Item) {
 	fmt.Printf("list\n")
 }
 
-func edit() {
-	// Create temp file
+func addItem(todos *[]Item) {
+
+}
+
+func editItem(todos *[]Item) {
+	usageInfo := "Usage: wtodo " + os.Args[1] + " <id>"
+
+	// Check for the ID command line argument
+	if len(os.Args) != 3 {
+		fmt.Fprintln(os.Stderr, usageInfo)
+		os.Exit(1)
+	}
+
+	// Find the element to edit
+	key, _ := strconv.Atoi(os.Args[2])
+	index := -1
+	for i := range *todos {
+		if (*todos)[i].Id == key {
+			index = i
+			break
+		}
+	}
+
+	// Exit if not found
+	if index == -1 {
+		fmt.Fprintln(os.Stderr, "ID not found:", os.Args[2], "\n", usageInfo)
+		list(*todos)
+		os.Exit(1)
+	}
+
+	// TODO: Add editing for the other options
+
+	// Create temp file to edit the name of the Item
 	temp, err := os.CreateTemp("", "tmp")
 	if err != nil {
 		log.Fatal(err)
 	}
 	defer temp.Close()
 	defer os.Remove(temp.Name())
+
+	// Write previous name to the file
+	temp.WriteString((*todos)[index].Name)
 
 	// Use $EDITOR as default editor or use vim
 	editor := os.Getenv("EDITOR")
@@ -39,10 +74,19 @@ func edit() {
 	}
 
 	// Read the file that the user wrote to
+	// and save that data to the array
 	content, err := ioutil.ReadFile(temp.Name())
 	if err != nil {
 		log.Fatal(err)
 	}
 	line := strings.Split(string(content), "\n")[0]
-	fmt.Println(line)
+	(*todos)[index].Name = line
+}
+
+func finishItem(todos *[]Item) {
+
+}
+
+func deleteItem(todos *[]Item) {
+
 }
